@@ -2,8 +2,8 @@ package com.ahmed.sampleApp.Activities;
 
 import android.app.Activity;
 import android.graphics.PorterDuff;
-import android.support.design.widget.TabLayout;
-import android.support.v4.view.ViewPager;
+import com.google.android.material.tabs.TabLayout;
+import androidx.viewpager.widget.ViewPager;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
